+++ pulsar/async/consts.py
@@ -1,4 +1,4 @@
-from pulsar.utils.structures import AttributeDictionary
+from pulsar.utils.structures import AttributeDictionary, FrozenDict
 
 # LOW LEVEL CONSTANTS - NO NEED TO CHANGE THOSE ###########################
 ACTOR_STATES = AttributeDictionary(INITIAL=0X0,
@@ -23,3 +23,12 @@ MIN_NOTIFY = 3     # DON'T NOTIFY BELOW THIS INTERVAL
 MAX_NOTIFY = 30    # NOTIFY AT LEAST AFTER THESE SECONDS
 ACTOR_TIMEOUT_TOLE = 0.3  # NOTIFY AFTER THIS TIMES THE TIMEOUT
 ACTOR_TERMINATE_TIMEOUT = 2 # TIMEOUT WHEN JOINING A TERMINATING ACTOR
+#
+# SPECIAL objects for Deferred
+CONTINUE = object()
+NOT_DONE = object()
+CLEAR_ERRORS = object()
+#
+# Globals
+EMPTY_TUPLE = ()
+EMPTY_DICT = FrozenDict()